agrisense-ai-main/model-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic

torch
torchvision
pillow
numpy

transformers
peft
accelerate
safetensors

ultralytics
